dom/browser-element/mochitest/priority/test_Audio.html

<!DOCTYPE HTML>
<html>
<!--
Test that frames playing audio get BACKGROUND_PERCEIVABLE priority.
-->
<head>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="../browserElementTestHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript;version=1.7">
"use strict";

SimpleTest.waitForExplicitFinish();
browserElementTestHelpers.setEnabledPref(true);
browserElementTestHelpers.addPermission();
browserElementTestHelpers.enableProcessPriorityManager();

// ProcessPriorityManager requires at least one process in foreground
// so that other processes can transit freely between foreground and
// background.
function setupTest() {
  var foreground = document.createElement('iframe');
  foreground.setAttribute('mozbrowser', true);
  foreground.src = browserElementTestHelpers.emptyPage;
  expectMozbrowserEvent(foreground, 'loadend').then(runTest);
  document.body.appendChild(foreground);
}

function runTest() {
  var iframe = document.createElement('iframe');
  iframe.setAttribute('mozbrowser', true);
  iframe.src = 'file_Audio.html';

  var childID = null;
  Promise.all([
    expectOnlyOneProcessCreated("FOREGROUND").then(function(chid) {
      childID = chid;
    }),
    expectMozbrowserEvent(iframe, 'loadend'),
    expectMozbrowserEvent(iframe, 'showmodalprompt').then(function(e) {
      is(e.detail.message, 'onplay', 'showmodalprompt message');
    })
  ]).then(function() {
    // Send the child process into the background.  Because it's playing audio,
    // it should get priority BACKGROUND_PERCEIVABLE, not vanilla BACKGROUND.
    var p = expectPriorityChange(childID, 'BACKGROUND_PERCEIVABLE');
    iframe.setVisible(false);
    return p;
  }).then(function() {
    var p = expectPriorityChange(childID, 'FOREGROUND');
    iframe.setVisible(true);
    return p;
  }).then(SimpleTest.finish);

  document.body.appendChild(iframe);
}

// This test relies on <audio> elements interacting with the audio channel
// service.  This is controled by the media.useAudioChannelService pref.
addEventListener('testready', function() {
  SpecialPowers.pushPrefEnv({set: [['media.useAudioChannelService', true]]},
                            setupTest);
});

</script>
</body>
</html>
